backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
scikit-learn
nltk
pandas
pydantic
python-dotenv
textblob
supabase
